requirements.txt
numpy>=1.26
pandas>=2.2
networkx>=3.2
tqdm>=4.66
matplotlib>=3.8
seaborn>=0.13
pyvis>=0.3.2
scipy>=1.12
polars>=0.20
python-igraph>=0.11
spacy[transformers]>=3.7
python-louvain>=0.16
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
cupy-cuda12x; platform_system=="Linux" or platform_system=="Windows"